api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
aiofiles==23.2.1
agno==0.1.0
google-generativeai==0.3.2
playwright==1.40.0
pillow==10.1.0
beautifulsoup4==4.12.2
requests==2.31.0
sqlalchemy==2.0.23
alembic==1.13.0
supabase==2.0.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4